requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.24.3
opencv-python-headless==4.8.1.78
pillow==10.1.0
python-telegram-bot==20.6
python-dotenv==1.0.0 